+++ src/compas_timber/elements/beam.py
@@ -14,13 +14,13 @@ from compas.geometry import bounding_box
 from compas.geometry import cross_vectors
 from compas.geometry import intersection_line_plane
 from compas.tolerance import TOL
-from compas_model.elements import Element
 from compas_model.elements import reset_computed
 
 from .features import FeatureApplicationError
+from .timber import TimberElement
 
 
-class Beam(Element):
+class Beam(TimberElement):
     """
     A class to represent timber beams (studs, slats, etc.) with rectangular cross-sections.
 
@@ -79,6 +79,8 @@ class Beam(Element):
         A list containing the 4 lines along the long axis of this beam.
     midpoint : :class:`~compas.geometry.Point`
         The point at the middle of the centerline of this beam.
+    key : int, optional
+        Once beam is added to a model, it will have this model-wide-unique integer key.
 
     """
 
@@ -110,6 +112,10 @@ class Beam(Element):
     # ==========================================================================
     # Computed attributes
     # ==========================================================================
+
+    @property
+    def is_beam(self):
+        return True
 
     @property
     def shape(self):
@@ -248,8 +254,13 @@ class Beam(Element):
 
     @property
     def has_features(self):
-        # TODO: move to compas_future... Part
+        # TODO: consider removing, this is not used anywhere
         return len(self.features) > 0
+
+    @property
+    def key(self):
+        # type: () -> int | None
+        return self.graph_node
 
     def __str__(self):
         return "Beam {:.3f} x {:.3f} x {:.3f} at {}".format(