+++ src/compas_timber/model/model.py
@@ -1,9 +1,12 @@
+import compas
+
+if not compas.IPY:
+    from typing import Generator  # noqa: F401
+
 from compas.geometry import Point
 from compas_model.models import Model
 
-from compas_timber.elements import Beam
-from compas_timber.elements import Plate
-from compas_timber.elements import Wall
+from compas_timber.connections import Joint
 
 
 class TimberModel(Model):
@@ -13,69 +16,70 @@ class TimberModel(Model):
 
     Attributes
     ----------
-    beams : list(:class:`~compas_timber.elements.Beam`)
-        A list of beams assigned to this model.
+    beams : Generator[:class:`~compas_timber.elements.Beam`]
+        A Generator object of all beams assigned to this model.
+    plates : Generator[:class:`~compas_timber.elements.Plate`]
+        A Generator object of all plates assigned to this model.
+    joints : Generator[:class:`~compas_timber.connections.Joint`]
+        A Generator object of all joints assigned to this model.
+    walls : Generator[:class:`~compas_timber.elements.Wall`]
+        A Generator object of all walls assigned to this model.
     center_of_mass : :class:`~compas.geometry.Point`
         The calculated center of mass of the model.
-    joints : list(:class:`~compas_timber.connections.Joint`)
-        A list of joints assigned to this model.
     topologies :  list(dict)
         A list of JointTopology for model. dict is: {"detected_topo": detected_topo, "beam_a_key": beam_a_key, "beam_b_key":beam_b_key}
         See :class:`~compas_timber.connections.JointTopology`.
     volume : float
         The calculated total volume of the model.
-    walls : list(:class:~compas_timber.elements.Wall)
-        A list of walls assigned to this model.
 
     """
 
     @classmethod
     def __from_data__(cls, data):
         model = super(TimberModel, cls).__from_data__(data)
-        for element in model.elements():
-            if isinstance(element, Beam):
-                model._beams.append(element)
-            elif isinstance(element, Plate):
-                model._plates.append(element)
-            elif isinstance(element, Wall):
-                model._walls.append(element)
         for interaction in model.interactions():
-            model._joints.append(interaction)
-            interaction.restore_beams_from_keys(model)
+            interaction.restore_beams_from_keys(model)  # type: ignore
         return model
 
     def __init__(self, *args, **kwargs):
         super(TimberModel, self).__init__()
-        self._beams = []
-        self._plates = []
-        self._walls = []
-        self._joints = []
         self._topologies = []  # added to avoid calculating multiple times
 
     def __str__(self):
-        return "TimberModel ({}) with {} beam(s), {} plate(s) and {} joint(s).".format(
-            self.guid, len(self.beams), len(self._plates), len(self.joints)
+        # type: () -> str
+        return "TimberModel ({}) with {} beam(s) and {} joint(s).".format(
+            str(self.guid), len(list(self.elements())), len(list(self.joints))
         )
 
     @property
     def beams(self):
-        # type: () -> list[Beam]
-        return self._beams
+        # type: () -> Generator[Beam, None, None]
+        # TODO: think about using `filter` instead of all these
+        # TODO: add `is_beam`, `is_plate` etc. to avoid using `isinstance`
+        for element in self.elements():
+            if getattr(element, "is_beam", False):
+                yield element
 
     @property
     def plates(self):
-        # type: () -> list[Plate]
-        return self._plates
+        # type: () -> Generator[Plate, None, None]
+        for element in self.elements():
+            if getattr(element, "is_plate", False):
+                yield element
 
     @property
     def joints(self):
-        # type: () -> list[Joint]
-        return self._joints
+        # type: () -> Generator[Joint, None, None]
+        for interaction in self.interactions():
+            if isinstance(interaction, Joint):
+                yield interaction  # TODO: consider if there are other interaction types...
 
     @property
     def walls(self):
-        # type: () -> list[Wall]
-        return self._walls
+        # type: () -> Generator[Wall, None, None]
+        for element in self.elements():
+            if getattr(element, "is_wall", False):
+                yield element
 
     @property
     def topologies(self):
@@ -87,10 +91,11 @@ class TimberModel(Model):
         total_vol = 0
         total_position = Point(0, 0, 0)
 
-        for beam in self._beams:
-            vol = beam.blank.volume
-            point = beam.blank_frame.point
-            point += beam.blank_frame.xaxis * (beam.blank_length / 2.0)
+        for element in self.elements():
+            vol = (
+                element.obb.volume
+            )  # TODO: include material density...? this uses volume as proxy for mass, which assumes all parts have equal density
+            point = element.obb.frame.point
             total_vol += vol
             total_position += point * vol
 
@@ -99,9 +104,9 @@ class TimberModel(Model):
     @property
     def volume(self):
         # type: () -> float
-        return sum([beam.blank.volume for beam in self._beams])  # TODO: add volume for plates
+        return sum([element.obb.volume for element in self.elements()])
 
-    def beam_by_guid(self, guid):
+    def element_by_guid(self, guid):
         # type: (str) -> Beam
         """Get a beam by its unique identifier.
 
@@ -112,26 +117,11 @@ class TimberModel(Model):
 
         Returns
         -------
-        :class:`~compas_timber.elements.Beam`
-            The beam with the specified GUID.
+        :class:`~compas_model.elements.Element`
+            The element with the specified GUID.
 
         """
         return self._guid_element[guid]
-
-    def add_element(self, element, **kwargs):
-        # TODO: make the distincion in the properties rather than here
-        # then get rid of this overrloading altogether.
-        node = super(TimberModel, self).add_element(element, **kwargs)
-
-        if isinstance(element, Beam):
-            self._beams.append(element)
-        elif isinstance(element, Wall):
-            self._walls.append(element)
-        elif isinstance(element, Plate):
-            self._plates.append(element)
-        else:
-            raise NotImplementedError("Element type not supported: {}".format(type(element)))
-        return node
 
     def add_joint(self, joint, beams):
         # type: (Joint, tuple[Beam]) -> None
@@ -150,7 +140,6 @@ class TimberModel(Model):
             raise ValueError("Expected 2 parts. Got instead: {}".format(len(beams)))
         a, b = beams
         _ = self.add_interaction(a, b, interaction=joint)
-        self._joints.append(joint)
 
     def remove_joint(self, joint):
         # type: (Joint) -> None
@@ -162,9 +151,8 @@ class TimberModel(Model):
             The joint to remove.
 
         """
-        a, b = joint.beams
-        self.remove_interaction(a, b)
-        self._joints.remove(joint)
+        a, b = joint.beams  # TODO: make this generic elements not beams
+        super(TimberModel, self).remove_interaction(a, b)  # TODO: Can two elements share more than one interaction?
 
     def set_topologies(self, topologies):
         """TODO: calculate the topologies inside the model using the ConnectionSolver."""