+++ tests/compas_timber/test_model.py
@@ -3,10 +3,13 @@ from compas.data import json_loads
 from compas.geometry import Frame
 from compas.geometry import Point
 from compas.geometry import Vector
+from compas.geometry import Polyline
 
 from compas_timber.connections import LButtJoint
 from compas_timber.connections import TButtJoint
 from compas_timber.elements import Beam
+from compas_timber.elements import Wall
+from compas_timber.elements import Plate
 from compas_timber.model import TimberModel
 
 
@@ -23,8 +26,8 @@ def test_add_element():
     assert B in A.beams
     assert len(list(A.graph.nodes())) == 1
     assert len(list(A.graph.edges())) == 0
-    assert A.beams[0] is B
-    assert len(A.beams) == 1
+    assert list(A.beams)[0] is B
+    assert len(list(A.beams)) == 1
 
 
 def test_add_joint():
@@ -36,8 +39,8 @@ def test_add_joint():
     model.add_element(b2)
     _ = LButtJoint.create(model, b1, b2)
 
-    assert len(model.beams) == 2
-    assert len(model.joints) == 1
+    assert len(list(model.beams)) == 2
+    assert len(list(model.joints)) == 1
 
 
 def test_copy(mocker):
@@ -53,7 +56,7 @@ def test_copy(mocker):
 
     A_copy = A.copy()
     assert A_copy is not A
-    assert A_copy.beams[0] is not A.beams[0]
+    assert list(A_copy.beams)[0] is not list(A.beams)[0]
 
 
 def test_deepcopy(mocker):
@@ -69,7 +72,7 @@ def test_deepcopy(mocker):
 
     A_copy = A.copy()
     assert A_copy is not A
-    assert A_copy.beams[0] is not A.beams[0]
+    assert list(A_copy.beams)[0] is not list(A.beams)[0]
 
 
 def test_beams_have_keys_after_serialization():
@@ -112,5 +115,59 @@ def test_serialization_with_t_butt_joints(mocker):
 
     a = json_loads(json_dumps(a))
 
-    assert len(a.joints) == 1
-    assert type(a.joints[0]) is TButtJoint
+    assert len(list(a.joints)) == 1
+    assert type(list(a.joints)[0]) is TButtJoint
+
+
+def test_generator_properties():
+    model = TimberModel()
+
+    polyline = Polyline(
+        [
+            Point(x=0.0, y=184.318671947, z=4252.92700512),
+            Point(x=0.0, y=2816.40294074, z=4252.92700512),
+            Point(x=0.0, y=2816.40294074, z=2720.97170805),
+            Point(x=0.0, y=184.318671947, z=2720.97170805),
+            Point(x=0.0, y=184.318671947, z=4252.92700512),
+        ]
+    )
+
+    plate = Plate(polyline, 10.0, Vector(1, 0, 0))
+    model.add_element(plate)
+
+    beam = Beam(Frame.worldXY(), 10.0, 10.0, 10.0)
+    model.add_element(beam)
+
+    wall = Wall(10.0, 10.0, 10.0, Frame.worldXY())
+    model.add_element(wall)
+
+    assert len(list(model.plates)) == 1
+    assert len(list(model.beams)) == 1
+    assert len(list(model.walls)) == 1
+
+
+def test_type_properties():
+    polyline = Polyline(
+        [
+            Point(x=0.0, y=184.318671947, z=4252.92700512),
+            Point(x=0.0, y=2816.40294074, z=4252.92700512),
+            Point(x=0.0, y=2816.40294074, z=2720.97170805),
+            Point(x=0.0, y=184.318671947, z=2720.97170805),
+            Point(x=0.0, y=184.318671947, z=4252.92700512),
+        ]
+    )
+
+    plate = Plate(polyline, 10.0, Vector(1, 0, 0))
+    beam = Beam(Frame.worldXY(), 10.0, 10.0, 10.0)
+    wall = Wall(10.0, 10.0, 10.0, Frame.worldXY())
+
+    assert plate.is_plate
+    assert beam.is_beam
+    assert wall.is_wall
+
+    assert not plate.is_beam
+    assert not plate.is_wall
+    assert not beam.is_wall
+    assert not beam.is_plate
+    assert not wall.is_plate
+    assert not wall.is_beam